requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
azure-core==1.26.3
azure-functions==1.13.2
azure-functions-durable==1.2.2
azure-identity==1.12.0
azure-storage-blob==12.15.0
cachetools==5.3.0
cffi==1.15.1
charset-normalizer==3.0.1
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
debugpy==1.6.6
fonttools==4.38.0
frozenlist==1.3.3
furl==2.1.3
google-api-core==2.11.0
google-auth==2.16.2
googleapis-common-protos==1.58.0
idna==3.4
importlib-resources==5.10.2
isodate==0.6.1
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
msal==1.21.0
msal-extensions==1.0.0
msrest==0.7.1
multidict==6.0.4
numpy==1.24.2
oauthlib==3.2.2
opencensus==0.11.1
opencensus-context==0.1.3
opencensus-ext-azure==1.1.8
opencensus-extension-azure-functions==1.0.0
orderedmultidict==1.0.1
packaging==23.0
Pillow==9.4.0
portalocker==2.7.0
protobuf==4.22.1
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.14
yarl==1.8.2
zipp==3.13.0
git+https://github.com/felihong/Azure-functionapps-demo.git